requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-ckeditor==6.7.1
django-colorfield==0.11.0
django-cors-headers==4.3.1
django-jalali==6.2.1
django-js-asset==2.2.0
django-render-partial==0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
idna==3.7
inflection==0.5.1
jalali_core==1.0.0
jdatetime==5.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mysqlclient==2.2.4
pillow==10.3.0
PyJWT==2.8.0
PyYAML==6.0.1
referencing==0.35.0
requests==2.31.0
rpds-py==0.18.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1